# Allow inlining methods with unmatched type parameters