@ianatol 2022-04-27
Re: partial reverting of #44512:

[`choltype`](https://github.com/JuliaLang/julia/blob/master/stdlib/LinearAlgebra/src/cholesky.jl#L181) from `LinearAlgebra` was causing BoundsErrors during test suite runs.

Looking further, it appeared that in the following IR,

```julia
Expr(:call, Main.eltype, Core.Argument(n=2)),
  Expr(:call, Main.oneunit, SSAValue(1)),
  Expr(:call, Main.sqrt, SSAValue(2)),
  Expr(:call, Main.typeof, SSAValue(3)),
  Expr(:call, Main.promote_type, SSAValue(4), Main.Float32),
  Core.ReturnNode(val=SSAValue(5))]
```

we were choosing to inline `promote_type` with the following match: `promote_type(DataType, Type{Float32}) from promote_type(Type{T}, Type{T}) where {T}`. 

This triggered `_compute_sparams` where
`m = promote_type(Type{T}, Type{T}) where {T}`
`tt = Tuple{typeof(Base.promote_type), Type{Float64}, Type{Float32}}`, which resulted in `env = svec()`, and our BoundsError.

Keno helped me to figure out that this was due to inlining not being able to discern which signature was right due to the `spec_types` of `promote_type(Type{T}, Type{T}) where {T}` and `promote_type(Type{T}, Type{S}) where {T,S}` being identical in our case. However, `DataType` in `spec_types` would later become `Float64`, meaning that this selection was actually incorrect. 

By bringing back [this branch](https://github.com/JuliaLang/julia/pull/45062/commits/b2ea6d114bfe7516bf40bd7e557e0e171c873864#diff-c08d0dffd36a142536203edece93dbe8656a68c607fd06256903d4b2ed8f2942R1284-R1310) that was removed in #44512, we were able to be a bit more fine-grained about when we could bypass validating sparams (as well as choosing not to inline when we had a direct match of `sig` and `spec_types`).