@ianatol 2022-04-28
`GeometricFlux`, `GraphSignals`, `GridapEmbedded`, `GridapGmsh`, `JetPack`, `NeuralOperators`, `Org`, ` PDENLPModels`, and `ValueShapes` are all caused by the same bug in FillArrays where we end up with `compute_sparams` on

`Tuple{Type{FillArrays.Fill{T, N, Axes} where Axes where N where T}, T, Vararg{Integer, N}} where N where T`
and
`Tuple{Type{FillArrays.Fill{T, N, Axes} where Axes where N where T}, Float32, Tuple{Int64}}`

Seems similar to the case listed above, but still investigating where a possibly incorrect inlining choice happens