@ianatol 2022-05-05
@aviatesk Could I ask for your review here? Specifically on [3a6b36c](https://github.com/JuliaLang/julia/pull/45062/commits/3a6b36c062c36c3250ee299cf7a3ef6672262782) where I revert some of #44512, as well as on [884e5ea](https://github.com/JuliaLang/julia/pull/45062/commits/884e5eaea0443f216b26bfe426b3a4121edca5a5), where I somewhat broadly restrict the cases in which we can perform this inlining. It feels like we could do better, but I'm not familiar enough with this part of the codebase to fully understand