@ianatol 2022-05-05
Re: test failures in `Metida` and `ODEInterface`:

The `ccall` and `cfunction`s showing up in the test aren't having their types properly calculated during inlining since we are entering `ssa_substitute_op!` while `spvals::SSAValue`. 

My idea is that we can do something similar to `svec_ref`, where we insert a node (in this case a call to `jl_instantiate_type_in_env`) where `spvals` is still an `SSAValue`. Then, before emitting the substituted `ccall`, we replace `spvals` in our inserted nodes with our calculation of the intersection.