@aviatesk 2022-05-06
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`