@aviatesk 2022-05-06
> By bringing back [this branch](https://github.com/JuliaLang/julia/pull/45062/commits/b2ea6d114bfe7516bf40bd7e557e0e171c873864#diff-c08d0dffd36a142536203edece93dbe8656a68c607fd06256903d4b2ed8f2942R1284-R1310) that was removed in #44512, we were able to be a bit more fine-grained about when we could bypass validating sparams (as well as choosing not to inline when we had a direct match of `sig` and `spec_types`).

#44512 removed the previous (somewhat random imho) special casings, and you can recover some if needed. But I don't want to revert it this way, i.e. which optimizable cases handled by #44512 are now ignored, and which are still optimized? I think you essentially need to compute `only_method` specifically, and it can co-exist with the code after #44512.