@nanosoldier 2022-05-06
[Your benchmark job](https://github.com/JuliaLang/julia/pull/45062#issuecomment-1119411463) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/0c08729_vs_902a5c1/report.md).