@nanosoldier 2022-07-14
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/45062#issuecomment-1183787515) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/e50ba3a_vs_415c668/report.html).