@ianatol 2022-08-09
Haven't been able to recreate the `SparseArrays/umfpack` error locally, so I guess it may be unrelated? Either way, this should be ready for another review when you get a chance! @aviatesk